requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
numpy==1.24.3
torch==1.13.1
faiss-cpu>=1.7.4
python-multipart>=0.0.6
imagebind @git+https://github.com/facebookresearch/ImageBind.git
prometheus-client>=0.17.0
sqlalchemy>=2.0.0
pymongo>=4.3.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.1
pandas>=2.1.4
pillow>=10.1.0
soundfile>=0.12.1
psutil
GPUtil
# torchvision==0.15.2
# torchaudio==2.0.2
# scikit-learn==1.3.0
# librosa==0.10.1
# transformers==4.35.0
# python-dotenv==1.0.0